Refactor ClassLoaderUtil into a new exported package.
