+++ build.gradle
@@ -490,6 +490,7 @@ project (':quasar-core') {
             from(zipTree(shadowJarTask.archiveFile)) {
                 exclude 'co/paralleluniverse/asm/**'
                 exclude 'co/paralleluniverse/common/asm/**'
+                exclude 'co/paralleluniverse/common/resource/**'
                 exclude 'co/paralleluniverse/fibers/instrument/**'
             }
             archiveAppendix = 'osgi'
@@ -534,6 +535,7 @@ Import-Package: \
             from(zipTree(shadowJarTask.archiveFile)) {
                 include 'co/paralleluniverse/asm/**'
                 include 'co/paralleluniverse/common/asm/**'
+                include 'co/paralleluniverse/common/resource/**'
                 include 'co/paralleluniverse/fibers/instrument/**'
                 exclude 'co/paralleluniverse/fibers/instrument/InstrumentationTask*'
                 exclude 'co/paralleluniverse/fibers/instrument/InstrumentationTKB*'
@@ -556,8 +558,8 @@ Bundle-Version: $version
             dependsOn bundleAgentTask
             doLast {
                 String importPackages = fetchManifestAttributeFrom(bundleAgentTask, 'Import-Package')
-                if (importPackages) {
-                    throw new InvalidUserDataException("OSGi bundle for Java agent should not import anything: " + importPackages)
+                if (importPackages && !importPackages.equals("co.paralleluniverse.common.resource")) {
+                    throw new InvalidUserDataException("OSGi bundle for Java agent should only import from itself: " + importPackages)
                 }
             }
         }
