+++ quasar-core/src/main/java/co/paralleluniverse/asm/ASMUtil.java
@@ -25,29 +25,13 @@ import java.io.InputStream;
 import java.util.List;
 import java.util.Objects;
 
+import static co.paralleluniverse.common.resource.ClassLoaderUtil.classToResource;
+
 /**
  *
  * @author pron
  */
 public final class ASMUtil {
-    public static String classToResource(String className) {
-        if (className == null)
-            return null;
-        return className.replace('.', '/') + ".class";
-    }
-
-    public static String classToSlashed(String className) {
-        if (className == null)
-            return null;
-        return className.replace('.', '/');
-    }
-
-    public static String classToSlashed(Class<?> clazz) {
-        if (clazz == null)
-            return null;
-        return classToSlashed(clazz.getName());
-    }
-
     public static InputStream getClassInputStream(String className, ClassLoader cl) {
         final String resource = classToResource(className);
         return cl != null ? cl.getResourceAsStream(resource) : ClassLoader.getSystemResourceAsStream(resource);
