+++ quasar-core/src/main/java/co/paralleluniverse/common/resource/ClassLoaderUtil.java
@@ -29,7 +29,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package co.paralleluniverse.common.reflection;
+package co.paralleluniverse.common.resource;
 
 import java.io.File;
 import java.io.IOException;
@@ -131,7 +131,7 @@ public final class ClassLoaderUtil {
     }
 
     private static void scanDirectory(File directory, ClassLoader classloader, Visitor visitor) throws IOException {
-        scanDirectory(directory, classloader, "", new HashSet<File>(), visitor);
+        scanDirectory(directory, classloader, "", new HashSet<>(), visitor);
     }
 
     private static void scanDirectory(File directory, ClassLoader classloader, String packagePrefix, Set<File> ancestors, Visitor visitor) throws IOException {
@@ -146,13 +146,12 @@ public final class ClassLoaderUtil {
             // IO error, just skip the directory
             return;
         }
-        Set<File> newAncestors = new HashSet<>();
-        newAncestors.addAll(ancestors);
+        Set<File> newAncestors = new HashSet<>(ancestors);
         newAncestors.add(canonical);
         for (File f : files) {
             String name = f.getName();
             if (f.isDirectory()) {
-                scanDirectory(f, classloader, packagePrefix + name + "/", newAncestors, visitor);
+                scanDirectory(f, classloader, packagePrefix + name + '/', newAncestors, visitor);
             } else {
                 String resourceName = packagePrefix + name;
                 if (!resourceName.equals(JarFile.MANIFEST_NAME))
